pass args to _psc_free()
git-svn-id: svn+ssh://frodo/cluster/svn/projects@13558 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ psc_fsutil_libs/tests/alloc/alloc_test.c
@@ -70,7 +70,7 @@ main(int argc, char *argv[])
 
 	p = psc_alloc(8, PAF_LOCK);
 	*(uint64_t *)p = 0;
-	psc_free(p, PAF_LOCK, 8);
+	psc_free(p, PAF_LOCK, (size_t)8);
 
 	sz = 1024;
 	p = psc_alloc(sz, PAF_LOCK | PAF_PAGEALIGN);
